Rename gp2dcp to dgp2dcp
