+++ cvxpy/reductions/dgp2dcp/atom_canonicalizers/__init__.py
@@ -15,22 +15,22 @@ from cvxpy.atoms.sum_largest import sum_largest
 from cvxpy.expressions.constants.constant import Constant
 from cvxpy.expressions.variable import Variable
 from cvxpy.reductions.eliminate_pwl.atom_canonicalizers import CANON_METHODS as PWL_METHODS
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.add_canon import add_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.constant_canon import constant_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.div_canon import div_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.exp_canon import exp_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.eye_minus_inv_canon import eye_minus_inv_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.geo_mean_canon import geo_mean_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.log_canon import log_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.mul_canon import mul_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.mulexpression_canon import mulexpression_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.nonpos_constr_canon import nonpos_constr_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.one_minus_canon import one_minus_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.pnorm_canon import pnorm_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.power_canon import power_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.spectral_radius_canon import spectral_radius_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.sum_canon import sum_canon
-from cvxpy.reductions.gp2dcp.atom_canonicalizers.trace_canon import trace_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.add_canon import add_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.constant_canon import constant_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.div_canon import div_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.exp_canon import exp_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.eye_minus_inv_canon import eye_minus_inv_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.geo_mean_canon import geo_mean_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.log_canon import log_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.mul_canon import mul_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.mulexpression_canon import mulexpression_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.nonpos_constr_canon import nonpos_constr_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.one_minus_canon import one_minus_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.pnorm_canon import pnorm_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.power_canon import power_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.spectral_radius_canon import spectral_radius_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.sum_canon import sum_canon
+from cvxpy.reductions.dgp2dcp.atom_canonicalizers.trace_canon import trace_canon
 
 
 # TODO(akshayka): canon for ...
