+++ cvxpy/reductions/dgp2dcp/atom_canonicalizers/constant_canon.py
