+++ cvxpy/reductions/dgp2dcp/atom_canonicalizers/geo_mean_canon.py
