+++ cvxpy/reductions/dgp2dcp/atom_canonicalizers/log_canon.py
