+++ cvxpy/reductions/dgp2dcp/atom_canonicalizers/mul_canon.py
