+++ cvxpy/reductions/dgp2dcp/atom_canonicalizers/mulexpression_canon.py
